Extend AssetsFrom handler from web.handlers.FilesFrom
This enables asynchronous handling as well as being able to use extended
headers, see https://github.com/xp-forge/web/pull/74#issuecomment-817117750

This pull request will bump the dependency on xp-forge/web to ^2.9 (and drop
support for version 1). Because BC is broken anyway, the with() function is
changed at the same time in backwards-incompatible way!